test(Subscription): Make AllowedFlightTime be a bigint with a converter to ticks
The default timespan type in the database is Time, which is not adequate
for a timespan because it reflects a date, with all date quirks including
arbitrary limitations on the maximum timespan that can be recorded. As
recommended by PomeloFoundation we should use a converter and force the
table to be bigint. See for details, including the link to Microsoft.

In this commit we needed to adapt the datatypes that convert minutes to
TimeSpan to be long, as the number of minutes in the maximum timespan can
be bigger than int.MaxValue. A test on the limits of AllowedFlightTime was
also added to make sure the actual TimeSpan limits map directly to the database
representation.

[1] https://github.com/PomeloFoundation/Pomelo.EntityFrameworkCore.MySql/issues/1046

+++ DroHub.Tests/TestInfrastructure/HttpClientHelper.cs
@@ -44,7 +44,7 @@ namespace DroHub.Tests.TestInfrastructure
                 string user_password,
                 string organization,
                 string user_base_type,
-                int allowed_flight_time_minutes,
+                long allowed_flight_time_minutes,
                 int allowed_user_count) {
 
                 if (user_email == "admin")
@@ -88,7 +88,7 @@ namespace DroHub.Tests.TestInfrastructure
                 string user_password,
                 string organization,
                 string user_base_type,
-                int allowed_flight_time_minutes,
+                long allowed_flight_time_minutes,
                 int allowed_user_count) {
 
                 return await addUser(test_fixture, "admin", test_fixture.AdminPassword, user_email, user_password,
